MAP-T dev checkpoint 5: Extended Port-Restricted Port Fields
Turns out the formal definition of the Port-Mapping Algorithm presented
in RFC 7597 is incomplete. The Port-Mapping Port Field is not supposed
to always length 16 bits; The actual length is 16 plus the suffix of the
"assigned prefix." When that happens, you have 16 bits worth of ports
for every address you're distributing through your CEs, hence it
multiplies. If the implementation doesn't take this into account, it
leaves a large portion of the port space unused (at least half) when
`o + r < 32`.

(If that explanation is too cryptic, try analyzing what happens to the
port distribution when `o + r < 32`. (See rfc7597#section-5.2.))

This commit reflects this discovery, patching `o + r < 32`.

------------------------------------------------------------------------

Unfortunately, this commit also became intermingled with a start of a
unit test framework refactor. Currently, onle the MAP-T unit test has
been upgraded. (But the other ones probably no longer work.)

The old unit tests were kernel modules that recompiled portions of Jool
to test them. There were two problems with this approach:

1. Each test yielded significant "portion" dependency management.
   (Because the dependency code of the tested code also needed to be
   tracked and recompiled for each unit test.)
2. Compiling everything over and over used to take too long.

The new unit tests will operate over the actual Jool kernel module, so
there will be only one dependency, and nothing will need to be
recompiled.

+++ src/common/config.h
@@ -502,13 +502,16 @@ struct port_restricted_port_field {
 	 * confusing if you're looking at the MAP IPv6 Address Format diagram.
 	 * (`a` does not have anything to do with `n + p`.)
 	 */
-	unsigned int a;
+	__u8 a;
 	/*
 	 * Length of the "PSID" field.
-	 * Seemingly also known as "q," unfortunately.
+	 * (Which is not the same as `q`.)
 	 */
-	unsigned int k;
-	/* "m" always equals 16 - a - k, so let's not store it. */
+	__u8 k;
+	/*
+	 * The length of `m` depends on `a`, `k` and the mapping rule in play.
+	 * Shouldn't store it.
+	 */
 };
 
 struct mapt_globals {
